Make OpgModel::type_name() return Cow instead
This change allow for dynamic type name generation without forcing an allocation when this is not needed.
Before this commit there was no way to return a dynamic type name
because the return type needed to be a statically allocated string. This
didn't allow to have generic structs with generated names. Consider the
following struct that represent a single page over a larger collection
of items:

```rust
struct Page<T> {
  items: Vec<T>,
}
```

We want its OpenAPI type to be dependent on T, so for example
`Page<User>` will have the type name `UserPage`.

Making `type_name()` return a `Cow<'static, str>` allows to implement it
like so:

```rust
fn type_name() -> Option<Cow<'static, str>> {
  match T::type_name() {
    Some(t) => Some(format!("{}Page", t).into()),
    None => Some("Page".into())
  }
}
```

If instead we wanted to use a static string, we can simply write:

```rust
fn type_name() -> Option<Cow<'static, str>> {
  Some("User".into())
}
```

`Cow<'string, str>` dereferences to `&str` so there is no
difference between using Cow or not for most usecases. To get ownership
of the type name we can simply call `name.into_owned()` and get back a
`String` instance. This will only clone, causing an allocation, if a
static name was returned. In case the `Cow` instead contained a
`String`, no allocation would be performed and the inner value would be
simply moved out of the `Cow`.

Signed-off-by: Matteo Joliveau <matteojoliveau@gmail.com>